Fix the normalization of Where definitions
