+++ dd-java-agent/agent-debugger/build.gradle
@@ -77,3 +77,6 @@ jar {
 // we want to test with no special reflective access (no --add-opens)
 ext.allowReflectiveAccessToJdk = false
 
+forbiddenApis {
+  ignoreFailures = true
+}
