+++ dd-java-agent/agent-debugger/src/test/java/com/datadog/debugger/agent/CapturedSnapshotTest.java
@@ -7,7 +7,6 @@ import static com.datadog.debugger.util.MoshiSnapshotHelper.NOT_CAPTURED_REASON;
 import static com.datadog.debugger.util.MoshiSnapshotHelper.REDACTED_IDENT_REASON;
 import static com.datadog.debugger.util.MoshiSnapshotHelper.REDACTED_TYPE_REASON;
 import static com.datadog.debugger.util.MoshiSnapshotTestHelper.VALUE_ADAPTER;
-import static com.datadog.debugger.util.TestHelper.setFieldInConfig;
 import static datadog.trace.bootstrap.debugger.util.Redaction.REDACTED_VALUE;
 import static org.junit.jupiter.api.Assertions.assertArrayEquals;
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -33,7 +32,6 @@ import com.datadog.debugger.el.values.StringValue;
 import com.datadog.debugger.instrumentation.InstrumentationResult;
 import com.datadog.debugger.probe.LogProbe;
 import com.datadog.debugger.probe.MetricProbe;
-import com.datadog.debugger.probe.ProbeDefinition;
 import com.datadog.debugger.probe.SpanDecorationProbe;
 import com.datadog.debugger.probe.SpanProbe;
 import com.datadog.debugger.probe.TriggerProbe;
@@ -41,12 +39,9 @@ import com.datadog.debugger.probe.Where;
 import com.datadog.debugger.sink.DebuggerSink;
 import com.datadog.debugger.sink.ProbeStatusSink;
 import com.datadog.debugger.sink.Snapshot;
-import com.datadog.debugger.util.MoshiHelper;
 import com.datadog.debugger.util.MoshiSnapshotTestHelper;
-import com.datadog.debugger.util.SerializerWithLimits;
 import com.datadog.debugger.util.TestSnapshotListener;
 import com.datadog.debugger.util.TestTraceInterceptor;
-import com.squareup.moshi.JsonAdapter;
 import datadog.trace.agent.tooling.TracerInstaller;
 import datadog.trace.api.Config;
 import datadog.trace.api.interceptor.MutableSpan;
@@ -67,37 +62,20 @@ import groovy.lang.GroovyClassLoader;
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.IOException;
-import java.lang.instrument.ClassFileTransformer;
-import java.lang.instrument.Instrumentation;
 import java.lang.reflect.Field;
 import java.lang.reflect.Modifier;
 import java.net.URISyntaxException;
 import java.net.URL;
-import java.net.URLClassLoader;
-import java.nio.file.Files;
-import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Collection;
 import java.util.Collections;
-import java.util.Comparator;
 import java.util.HashMap;
-import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
-import net.bytebuddy.agent.ByteBuddyAgent;
-import org.jetbrains.kotlin.cli.common.ExitCode;
-import org.jetbrains.kotlin.cli.common.arguments.K2JVMCompilerArguments;
-import org.jetbrains.kotlin.cli.common.messages.MessageRenderer;
-import org.jetbrains.kotlin.cli.common.messages.PrintingMessageCollector;
-import org.jetbrains.kotlin.cli.jvm.K2JVMCompiler;
-import org.jetbrains.kotlin.config.Services;
 import org.joor.Reflect;
 import org.joor.ReflectException;
-import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.condition.DisabledIf;
@@ -109,36 +87,11 @@ import org.junit.jupiter.params.provider.ValueSource;
 import org.mockito.ArgumentCaptor;
 import utils.SourceCompiler;
 
-public class CapturedSnapshotTest {
-  private static final String LANGUAGE = "java";
-  private static final ProbeId PROBE_ID = new ProbeId("beae1807-f3b0-4ea8-a74f-826790c5e6f5", 0);
+public class CapturedSnapshotTest extends CapturingTestBase {
   private static final ProbeId PROBE_ID1 = new ProbeId("beae1807-f3b0-4ea8-a74f-826790c5e6f6", 0);
   private static final ProbeId PROBE_ID2 = new ProbeId("beae1807-f3b0-4ea8-a74f-826790c5e6f7", 0);
   private static final ProbeId PROBE_ID3 = new ProbeId("beae1807-f3b0-4ea8-a74f-826790c5e6f8", 0);
   private static final ProbeId PROBE_ID4 = new ProbeId("beae1807-f3b0-4ea8-a74f-826790c5e6f9", 0);
-  private static final String SERVICE_NAME = "service-name";
-  private static final JsonAdapter<Map<String, Object>> GENERIC_ADAPTER =
-      MoshiHelper.createGenericAdapter();
-
-  private Instrumentation instr = ByteBuddyAgent.install();
-  private ClassFileTransformer currentTransformer;
-  private ProbeStatusSink probeStatusSink;
-  private MockInstrumentationListener instrumentationListener;
-
-  @BeforeEach
-  public void before() {
-    setFieldInConfig(Config.get(), "debuggerCaptureTimeout", 200);
-  }
-
-  @AfterEach
-  public void after() {
-    if (currentTransformer != null) {
-      instr.removeTransformer(currentTransformer);
-    }
-    ProbeRateLimiter.resetAll();
-    Assertions.assertFalse(DebuggerContext.isInProbe());
-    Redaction.clearUserDefinedTypes();
-  }
 
   @Test
   public void bridgeMethods() throws IOException, URISyntaxException {
@@ -215,7 +168,7 @@ public class CapturedSnapshotTest {
     final String CLASS_NAME = "CapturedSnapshot01";
     LogProbe lineProbe = createProbe(PROBE_ID1, CLASS_NAME, "main", "int (java.lang.String)", "8");
     LogProbe methodProbe = createProbe(PROBE_ID2, CLASS_NAME, "main", "int (java.lang.String)");
-    TestSnapshotListener listener = installProbes(CLASS_NAME, lineProbe, methodProbe);
+    TestSnapshotListener listener = installProbes(lineProbe, methodProbe);
     DebuggerAgentHelper.injectSink(listener);
     DebuggerContext.initProbeResolver(
         (encodedProbeId) -> {
@@ -322,7 +275,6 @@ public class CapturedSnapshotTest {
     final String CLASS_NAME = "com.datadog.debugger.CapturedSnapshot15";
     TestSnapshotListener listener =
         installProbes(
-            CLASS_NAME,
             createProbe(PROBE_ID1, CLASS_NAME, "<init>", "()"),
             createProbe(PROBE_ID2, CLASS_NAME, "<init>", "(String, long, String)"));
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
@@ -336,7 +288,6 @@ public class CapturedSnapshotTest {
     final String CLASS_NAME = "CapturedSnapshot03";
     TestSnapshotListener listener =
         installProbes(
-            CLASS_NAME,
             createProbe(PROBE_ID1, CLASS_NAME, "f1", "(int)"),
             createProbe(PROBE_ID2, CLASS_NAME, "f2", "(int)"));
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
@@ -356,7 +307,7 @@ public class CapturedSnapshotTest {
     final String CLASS_NAME = "CapturedSnapshot03";
     LogProbe probe = createProbe(PROBE_ID1, CLASS_NAME, "f1", "(int)");
     LogProbe probe2 = createProbe(PROBE_ID2, CLASS_NAME, "f1", "(int)");
-    TestSnapshotListener listener = installProbes(CLASS_NAME, probe, probe2);
+    TestSnapshotListener listener = installProbes(probe, probe2);
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
     int result = Reflect.onClass(testClass).call("main", "").get();
     assertEquals(48, result);
@@ -381,8 +332,7 @@ public class CapturedSnapshotTest {
   @Test
   public void catchBlock() throws IOException, URISyntaxException {
     final String CLASS_NAME = "CapturedSnapshot02";
-    TestSnapshotListener listener =
-        installProbes(CLASS_NAME, createProbe(PROBE_ID, CLASS_NAME, "f", "()"));
+    TestSnapshotListener listener = installProbes(createProbe(PROBE_ID, CLASS_NAME, "f", "()"));
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
     int result = Reflect.onClass(testClass).call("main", "f").get();
     assertEquals(42, result);
@@ -397,7 +347,6 @@ public class CapturedSnapshotTest {
     final int LINE_END = 48;
     TestSnapshotListener listener =
         installProbes(
-            CLASS_NAME,
             createProbe(
                 PROBE_ID, CLASS_NAME, "synchronizedBlock", "(int)", LINE_START + "-" + LINE_END));
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
@@ -425,7 +374,6 @@ public class CapturedSnapshotTest {
     final int LINE_END = 49;
     TestSnapshotListener listener =
         installProbes(
-            CLASS_NAME,
             createProbe(
                 PROBE_ID, CLASS_NAME, "synchronizedBlock", "(int)", LINE_START + "-" + LINE_END));
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
@@ -443,7 +391,7 @@ public class CapturedSnapshotTest {
   public void sourceFileProbe() throws IOException, URISyntaxException {
     final String CLASS_NAME = "CapturedSnapshot03";
     TestSnapshotListener listener =
-        installProbes(CLASS_NAME, createSourceFileProbe(PROBE_ID, CLASS_NAME + ".java", 4));
+        installProbes(createSourceFileProbe(PROBE_ID, CLASS_NAME + ".java", 4));
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
     int result = Reflect.onClass(testClass).call("main", "").get();
     assertEquals(48, result);
@@ -460,7 +408,7 @@ public class CapturedSnapshotTest {
   public void simpleSourceFileProbe() throws IOException, URISyntaxException {
     final String CLASS_NAME = "com.datadog.debugger.CapturedSnapshot10";
     TestSnapshotListener listener =
-        installProbes(CLASS_NAME, createSourceFileProbe(PROBE_ID, "CapturedSnapshot10.java", 11));
+        installProbes(createSourceFileProbe(PROBE_ID, "CapturedSnapshot10.java", 11));
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
     int result = Reflect.onClass(testClass).call("main", "2").get();
     assertEquals(2, result);
@@ -479,7 +427,6 @@ public class CapturedSnapshotTest {
     String DIR_CLASS_NAME = CLASS_NAME.replace('.', '/');
     TestSnapshotListener listener =
         installProbes(
-            CLASS_NAME,
             createSourceFileProbe(PROBE_ID, "src/main/java/" + DIR_CLASS_NAME + ".java", 11));
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
     int result = Reflect.onClass(testClass).call("main", "2").get();
@@ -499,7 +446,6 @@ public class CapturedSnapshotTest {
     String DIR_CLASS_NAME = CLASS_NAME.replace('.', '/');
     TestSnapshotListener listener =
         installProbes(
-            "com.datadog.debugger.TopLevel01",
             createSourceFileProbe(PROBE_ID, "src/main/java/" + DIR_CLASS_NAME + ".java", 21));
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
     int result = Reflect.onClass(testClass).call("main", "1").get();
@@ -520,7 +466,6 @@ public class CapturedSnapshotTest {
     String DIR_CLASS_NAME = CLASS_NAME.replace('.', '/');
     TestSnapshotListener listener =
         installProbes(
-            CLASS_NAME,
             createSourceFileProbe(PROBE_ID1, "src/main/java/" + DIR_CLASS_NAME + ".java", 10),
             createProbe(PROBE_ID2, CLASS_NAME, "main", null));
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
@@ -548,8 +493,7 @@ public class CapturedSnapshotTest {
   public void sourceFileProbeScala() throws IOException, URISyntaxException {
     final String CLASS_NAME = "CapturedSnapshot101";
     final String FILE_NAME = CLASS_NAME + ".scala";
-    TestSnapshotListener listener =
-        installProbes(CLASS_NAME, createSourceFileProbe(PROBE_ID, FILE_NAME, 3));
+    TestSnapshotListener listener = installProbes(createSourceFileProbe(PROBE_ID, FILE_NAME, 3));
     String source = getFixtureContent("/" + FILE_NAME);
     Class<?> testClass = ScalaHelper.compileAndLoad(source, CLASS_NAME, FILE_NAME);
     int result = Reflect.onClass(testClass).call("main", "").get();
@@ -567,7 +511,7 @@ public class CapturedSnapshotTest {
   public void sourceFileProbeGroovy() throws IOException, URISyntaxException {
     final String CLASS_NAME = "CapturedSnapshot201";
     TestSnapshotListener listener =
-        installProbes(CLASS_NAME, createSourceFileProbe(PROBE_ID, CLASS_NAME + ".groovy", 4));
+        installProbes(createSourceFileProbe(PROBE_ID, CLASS_NAME + ".groovy", 4));
     String source = getFixtureContent("/" + CLASS_NAME + ".groovy");
     GroovyClassLoader groovyClassLoader = new GroovyClassLoader();
     Class<?> testClass = groovyClassLoader.parseClass(source);
@@ -589,7 +533,7 @@ public class CapturedSnapshotTest {
   public void sourceFileProbeKotlin() {
     final String CLASS_NAME = "CapturedSnapshot301";
     TestSnapshotListener listener =
-        installProbes(CLASS_NAME, createSourceFileProbe(PROBE_ID, CLASS_NAME + ".kt", 4));
+        installProbes(createSourceFileProbe(PROBE_ID, CLASS_NAME + ".kt", 4));
     URL resource = CapturedSnapshotTest.class.getResource("/" + CLASS_NAME + ".kt");
     assertNotNull(resource);
     List<File> filesToDelete = new ArrayList<>();
@@ -668,7 +612,7 @@ public class CapturedSnapshotTest {
     LogProbe simpleDataProbe = builder.capture(1, 100, 255, Limits.DEFAULT_FIELD_COUNT).build();
     builder = createProbeBuilder(PROBE_ID2, CLASS_NAME, "createCompositeData", "()");
     LogProbe compositeDataProbe = builder.capture(1, 3, 255, Limits.DEFAULT_FIELD_COUNT).build();
-    TestSnapshotListener listener = installProbes(CLASS_NAME, simpleDataProbe, compositeDataProbe);
+    TestSnapshotListener listener = installProbes(simpleDataProbe, compositeDataProbe);
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
     int result = Reflect.onClass(testClass).call("main", "").get();
     assertEquals(143, result);
@@ -699,7 +643,7 @@ public class CapturedSnapshotTest {
     LogProbe.Builder builder =
         createProbeBuilder(PROBE_ID, CLASS_NAME, "createCompositeData", "()");
     LogProbe compositeDataProbe = builder.capture(2, 3, 255, Limits.DEFAULT_FIELD_COUNT).build();
-    TestSnapshotListener listener = installProbes(CLASS_NAME, compositeDataProbe);
+    TestSnapshotListener listener = installProbes(compositeDataProbe);
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
     int result = Reflect.onClass(testClass).call("main", "").get();
     assertEquals(143, result);
@@ -723,7 +667,7 @@ public class CapturedSnapshotTest {
     final String CLASS_NAME = "CapturedSnapshot04";
     LogProbe.Builder builder = createProbeBuilder(PROBE_ID, CLASS_NAME, "createSimpleData", "()");
     LogProbe simpleDataProbe = builder.capture(1, 100, 2, Limits.DEFAULT_FIELD_COUNT).build();
-    TestSnapshotListener listener = installProbes(CLASS_NAME, simpleDataProbe);
+    TestSnapshotListener listener = installProbes(simpleDataProbe);
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
     int result = Reflect.onClass(testClass).call("main", "").get();
     assertEquals(143, result);
@@ -741,7 +685,7 @@ public class CapturedSnapshotTest {
     final String CLASS_NAME = "CapturedSnapshot04";
     LogProbe.Builder builder = createProbeBuilder(PROBE_ID, CLASS_NAME, "createSimpleData", "()");
     LogProbe simpleDataProbe = builder.capture(0, 100, 50, Limits.DEFAULT_FIELD_COUNT).build();
-    TestSnapshotListener listener = installProbes(CLASS_NAME, simpleDataProbe);
+    TestSnapshotListener listener = installProbes(simpleDataProbe);
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
     int result = Reflect.onClass(testClass).call("main", "").get();
     assertEquals(143, result);
@@ -758,7 +702,7 @@ public class CapturedSnapshotTest {
     final String CLASS_NAME = "CapturedSnapshot04";
     LogProbe.Builder builder = createProbeBuilder(PROBE_ID, CLASS_NAME, "createSimpleData", "()");
     LogProbe simpleDataProbe = builder.capture(0, 100, 50, Limits.DEFAULT_FIELD_COUNT).build();
-    TestSnapshotListener listener = installProbes(CLASS_NAME, simpleDataProbe);
+    TestSnapshotListener listener = installProbes(simpleDataProbe);
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
     int result = Reflect.onClass(testClass).call("main", "").get();
     assertEquals(143, result);
@@ -776,7 +720,7 @@ public class CapturedSnapshotTest {
     final String CLASS_NAME = "CapturedSnapshot04";
     LogProbe.Builder builder = createProbeBuilder(PROBE_ID, CLASS_NAME, "createSimpleData", "()");
     LogProbe simpleDataProbe = builder.capture(1, 100, 50, Limits.DEFAULT_FIELD_COUNT).build();
-    TestSnapshotListener listener = installProbes(CLASS_NAME, simpleDataProbe);
+    TestSnapshotListener listener = installProbes(simpleDataProbe);
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
     int result = Reflect.onClass(testClass).call("main", "").get();
     assertEquals(143, result);
@@ -796,7 +740,7 @@ public class CapturedSnapshotTest {
     LogProbe.Builder builder = createProbeBuilder(PROBE_ID, CLASS_NAME, "createSimpleData", "()");
     LogProbe probe1 = builder.capture(0, 100, 50, Limits.DEFAULT_FIELD_COUNT).build();
     LogProbe probe2 = builder.capture(3, 100, 50, Limits.DEFAULT_FIELD_COUNT).build();
-    TestSnapshotListener listener = installProbes(CLASS_NAME, probe1, probe2);
+    TestSnapshotListener listener = installProbes(probe1, probe2);
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
     int result = Reflect.onClass(testClass).call("main", "").get();
     assertEquals(143, result);
@@ -815,7 +759,7 @@ public class CapturedSnapshotTest {
     LogProbe.Builder builder =
         createProbeBuilder(PROBE_ID, CLASS_NAME, "createCompositeData", "()");
     LogProbe compositeDataProbe = builder.capture(2, 3, 255, 2).build();
-    TestSnapshotListener listener = installProbes(CLASS_NAME, compositeDataProbe);
+    TestSnapshotListener listener = installProbes(compositeDataProbe);
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
     int result = Reflect.onClass(testClass).call("main", "").get();
     assertEquals(143, result);
@@ -851,7 +795,7 @@ public class CapturedSnapshotTest {
   public void fieldExtractorNotAccessible() throws IOException, URISyntaxException {
     final String CLASS_NAME = "com.datadog.debugger.CapturedSnapshot30";
     LogProbe logProbe = createProbe(PROBE_ID, CLASS_NAME + "$MyObjectInputStream", "process", "()");
-    TestSnapshotListener listener = installProbes(CLASS_NAME, logProbe);
+    TestSnapshotListener listener = installProbes(logProbe);
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
     int result = Reflect.onClass(testClass).call("main", "").get();
     assertEquals(42, result);
@@ -871,7 +815,7 @@ public class CapturedSnapshotTest {
   public void staticFieldExtractorNotAccessible() throws IOException, URISyntaxException {
     final String CLASS_NAME = "com.datadog.debugger.CapturedSnapshot30";
     LogProbe logProbe = createProbe(PROBE_ID, CLASS_NAME + "$MyHttpURLConnection", "process", "()");
-    TestSnapshotListener listener = installProbes(CLASS_NAME, logProbe);
+    TestSnapshotListener listener = installProbes(logProbe);
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
     int result = Reflect.onClass(testClass).call("main", "static").get();
     assertEquals(42, result);
@@ -890,8 +834,7 @@ public class CapturedSnapshotTest {
   public void uncaughtException() throws IOException, URISyntaxException {
     final String CLASS_NAME = "CapturedSnapshot05";
     TestSnapshotListener listener =
-        installProbes(
-            CLASS_NAME, createProbe(PROBE_ID, CLASS_NAME, "triggerUncaughtException", "()"));
+        installProbes(createProbe(PROBE_ID, CLASS_NAME, "triggerUncaughtException", "()"));
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
     try {
       Reflect.onClass(testClass).call("main", "triggerUncaughtException").get();
@@ -939,7 +882,7 @@ public class CapturedSnapshotTest {
                     "@exception instanceof \"CapturedSnapshot05$CustomException\" and @exception.detailMessage == 'oops' and @exception.additionalMsg == 'I did it again'"))
             .template(LOG_TEMPLATE, parseTemplate(LOG_TEMPLATE))
             .build();
-    TestSnapshotListener listener = installProbes(CLASS_NAME, probe);
+    TestSnapshotListener listener = installProbes(probe);
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
     try {
       Reflect.onClass(testClass).call("main", "triggerUncaughtException").get();
@@ -961,8 +904,7 @@ public class CapturedSnapshotTest {
   public void caughtException() throws IOException, URISyntaxException {
     final String CLASS_NAME = "CapturedSnapshot05";
     TestSnapshotListener listener =
-        installProbes(
-            CLASS_NAME, createProbe(PROBE_ID, CLASS_NAME, "triggerCaughtException", "()"));
+        installProbes(createProbe(PROBE_ID, CLASS_NAME, "triggerCaughtException", "()"));
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
     int result = Reflect.onClass(testClass).call("main", "triggerCaughtException").get();
     assertEquals(42, result);
@@ -981,7 +923,7 @@ public class CapturedSnapshotTest {
     LogProbe probe1 = createProbe(PROBE_ID1, CLASS_NAME, "triggerSwallowedException", null, "26");
     LogProbe probe2 = createProbe(PROBE_ID2, CLASS_NAME, "triggerSwallowedException", null, "29");
     LogProbe probe3 = createProbe(PROBE_ID3, CLASS_NAME, "triggerSwallowedException", null, "32");
-    TestSnapshotListener listener = installProbes(CLASS_NAME, probe1, probe2, probe3);
+    TestSnapshotListener listener = installProbes(probe1, probe2, probe3);
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
     int result = Reflect.onClass(testClass).call("main", "triggerSwallowedException").get();
     assertEquals(-1, result);
@@ -1025,7 +967,7 @@ public class CapturedSnapshotTest {
                     "not(isDefined(@exception))"))
             .template(LOG_TEMPLATE, parseTemplate(LOG_TEMPLATE))
             .build();
-    TestSnapshotListener listener = installProbes(CLASS_NAME, probe);
+    TestSnapshotListener listener = installProbes(probe);
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
     int result = Reflect.onClass(testClass).call("main", "2").get();
     assertEquals(2, result);
@@ -1042,7 +984,7 @@ public class CapturedSnapshotTest {
             .where(CLASS_NAME, 8)
             .sampling(new LogProbe.Sampling(1))
             .build();
-    TestSnapshotListener listener = installProbes(CLASS_NAME, logProbes);
+    TestSnapshotListener listener = installProbes(logProbes);
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
     for (int i = 0; i < 100; i++) {
       int result = Reflect.onClass(testClass).call("main", "1").get();
@@ -1062,7 +1004,7 @@ public class CapturedSnapshotTest {
             .addLogProbes(Arrays.asList(probe1, probe2))
             .add(new LogProbe.Sampling(1))
             .build();
-    TestSnapshotListener listener = installProbes(CLASS_NAME, config);
+    TestSnapshotListener listener = installProbes(config);
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
     for (int i = 0; i < 100; i++) {
       int result = Reflect.onClass(testClass).call("main", "").get();
@@ -1097,7 +1039,7 @@ public class CapturedSnapshotTest {
                     "(fld == 11 && typed.fld.fld.msg == \"hello\") && (arg == '5' && @duration >= 0)"))
             .evaluateAt(MethodLocation.EXIT)
             .build();
-    TestSnapshotListener listener = installProbes(CLASS_NAME, logProbe);
+    TestSnapshotListener listener = installProbes(logProbe);
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
     for (int i = 0; i < 100; i++) {
       int result = Reflect.onClass(testClass).call("main", String.valueOf(i)).get();
@@ -1131,7 +1073,7 @@ public class CapturedSnapshotTest {
                             DSL.eq(DSL.ref("arg"), DSL.value("5")))),
                     "(fld == 11 && typed.fld.fld.msg == \"hello\") && arg == '5'"))
             .build();
-    TestSnapshotListener listener = installProbes(CLASS_NAME, logProbe);
+    TestSnapshotListener listener = installProbes(logProbe);
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
     for (int i = 0; i < 100; i++) {
       int result = Reflect.onClass(testClass).call("main", String.valueOf(i)).get();
@@ -1152,7 +1094,7 @@ public class CapturedSnapshotTest {
                     DSL.when(DSL.eq(DSL.ref("strField"), DSL.value("foo"))), "strField == 'foo'"))
             .evaluateAt(MethodLocation.EXIT)
             .build();
-    TestSnapshotListener listener = installProbes(CLASS_NAME, logProbe);
+    TestSnapshotListener listener = installProbes(logProbe);
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
     int result = Reflect.onClass(testClass).call("main", "0").get();
     assertEquals(42, result);
@@ -1176,7 +1118,7 @@ public class CapturedSnapshotTest {
                 new ProbeCondition(DSL.when(DSL.eq(DSL.ref("arg"), DSL.value("5"))), "arg == '5'"))
             .evaluateAt(MethodLocation.EXIT)
             .build();
-    TestSnapshotListener listener = installProbes(CLASS_NAME, logProbe);
+    TestSnapshotListener listener = installProbes(logProbe);
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
     int result = Reflect.onClass(testClass).call("main", "0").get();
     assertEquals(3, result);
@@ -1198,7 +1140,7 @@ public class CapturedSnapshotTest {
                             DSL.value("hello"))),
                     "nullTyped.fld.fld.msg == 'hello'"))
             .build();
-    TestSnapshotListener listener = installProbes(CLASS_NAME, logProbes);
+    TestSnapshotListener listener = installProbes(logProbes);
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
     int result = Reflect.onClass(testClass).call("main", "1").get();
     Snapshot snapshot = assertOneSnapshot(listener);
@@ -1224,7 +1166,7 @@ public class CapturedSnapshotTest {
                                 new StringValue("closed")))),
                     "isDefined(@exception) && contains(@exception.detailMessage, 'closed')"))
             .build();
-    TestSnapshotListener listener = installProbes(CLASS_NAME, logProbes);
+    TestSnapshotListener listener = installProbes(logProbes);
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
     int result = Reflect.onClass(testClass).call("main", "1").get();
     assertEquals(3, result);
@@ -1244,7 +1186,7 @@ public class CapturedSnapshotTest {
                             DSL.getMember(DSL.ref("maybeStr"), "value"), DSL.value("maybe foo"))),
                     "maybeStr.value == 'maybe foo'"))
             .build();
-    TestSnapshotListener listener = installProbes(CLASS_NAME, logProbes);
+    TestSnapshotListener listener = installProbes(logProbes);
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
     int result = Reflect.onClass(testClass).call("main", "1").get();
     assertEquals(3, result);
@@ -1308,7 +1250,7 @@ public class CapturedSnapshotTest {
         createProbeBuilder(PROBE_ID2, CLASS_NAME, "doit", "int (java.lang.String)")
             .when(probeCondition2)
             .build();
-    TestSnapshotListener listener = installProbes(CLASS_NAME, probe1, probe2);
+    TestSnapshotListener listener = installProbes(probe1, probe2);
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
     int result = Reflect.onClass(testClass).call("main", "1").get();
     assertEquals(3, result);
@@ -1409,7 +1351,7 @@ public class CapturedSnapshotTest {
                     DSL.when(DSL.ge(DSL.ref("@duration"), DSL.value(0))), "@duration >= 0"))
             .evaluateAt(MethodLocation.EXIT)
             .build();
-    TestSnapshotListener listener = installProbes(CLASS_NAME, probe1, probe2);
+    TestSnapshotListener listener = installProbes(probe1, probe2);
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
     int result = Reflect.onClass(testClass).call("main", "1").get();
     Assertions.assertEquals(3, result);
@@ -1421,8 +1363,7 @@ public class CapturedSnapshotTest {
   @Test
   public void fields() throws IOException, URISyntaxException {
     final String CLASS_NAME = "CapturedSnapshot06";
-    TestSnapshotListener listener =
-        installProbes(CLASS_NAME, createProbe(PROBE_ID, CLASS_NAME, "f", "()"));
+    TestSnapshotListener listener = installProbes(createProbe(PROBE_ID, CLASS_NAME, "f", "()"));
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
     int result = Reflect.onClass(testClass).call("main", "f").get();
     assertEquals(42, result);
@@ -1462,7 +1403,7 @@ public class CapturedSnapshotTest {
                     DSL.when(DSL.eq(DSL.ref("intValue"), DSL.value(24))), "intValue == 24"))
             .evaluateAt(MethodLocation.ENTRY)
             .build();
-    TestSnapshotListener listener = installProbes(INHERITED_CLASS_NAME, probe);
+    TestSnapshotListener listener = installProbes(probe);
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
     int result = Reflect.onClass(testClass).call("main", "inherited").get();
     assertEquals(42, result);
@@ -1505,7 +1446,7 @@ public class CapturedSnapshotTest {
                     DSL.when(DSL.eq(DSL.ref("intValue"), DSL.value(48))), "intValue == 48"))
             .evaluateAt(MethodLocation.EXIT)
             .build();
-    TestSnapshotListener listener = installProbes(INHERITED_CLASS_NAME, logProbe);
+    TestSnapshotListener listener = installProbes(logProbe);
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
     int result = Reflect.onClass(testClass).call("main", "inherited").get();
     assertEquals(42, result);
@@ -1527,7 +1468,7 @@ public class CapturedSnapshotTest {
     setCorrelationSingleton(spyCorrelationAccess);
     doReturn(true).when(spyCorrelationAccess).isAvailable();
     TestSnapshotListener listener =
-        installProbes(CLASS_NAME, createProbe(PROBE_ID, CLASS_NAME, null, null, "37"));
+        installProbes(createProbe(PROBE_ID, CLASS_NAME, null, null, "37"));
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
     int result = Reflect.onClass(testClass).call("main", "static", "email@address").get();
     assertEquals(8, result);
@@ -1544,7 +1485,7 @@ public class CapturedSnapshotTest {
     setCorrelationSingleton(spyCorrelationAccess);
     doReturn(true).when(spyCorrelationAccess).isAvailable();
     TestSnapshotListener listener =
-        installProbes(CLASS_NAME, createProbe(PROBE_ID, CLASS_NAME, null, null, "48"));
+        installProbes(createProbe(PROBE_ID, CLASS_NAME, null, null, "48"));
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
     int result = Reflect.onClass(testClass).call("main", "capturing", "email@address").get();
     assertEquals(8, result);
@@ -1562,7 +1503,7 @@ public class CapturedSnapshotTest {
     setCorrelationSingleton(spyCorrelationAccess);
     doReturn(true).when(spyCorrelationAccess).isAvailable();
     TestSnapshotListener listener =
-        installProbes(CLASS_NAME, createProbe(PROBE_ID, CLASS_NAME, null, null, "58"));
+        installProbes(createProbe(PROBE_ID, CLASS_NAME, null, null, "58"));
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
     int result = Reflect.onClass(testClass).call("main", "multi", "FOO1,FOO2,FOO3").get();
     assertEquals(3, result);
@@ -1603,8 +1544,7 @@ public class CapturedSnapshotTest {
     final String CLASS_NAME = "CapturedSnapshot09";
     LogProbe nativeMethodProbe = createProbe(PROBE_ID1, CLASS_NAME, "nativeMethod", "()");
     LogProbe abstractMethodProbe = createProbe(PROBE_ID2, CLASS_NAME, "abstractMethod", "()");
-    TestSnapshotListener listener =
-        installProbes(CLASS_NAME, nativeMethodProbe, abstractMethodProbe);
+    TestSnapshotListener listener = installProbes(nativeMethodProbe, abstractMethodProbe);
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
     int result = Reflect.onClass(testClass).call("main", "").get();
     assertEquals(1, result);
@@ -1626,7 +1566,7 @@ public class CapturedSnapshotTest {
     // @1f7ef9ea, parent loader 'app')
     final String CLASS_NAME = "com.datadog.debugger.CapturedSnapshot12";
     LogProbe abstractMethodProbe = createProbe(PROBE_ID, CLASS_NAME, "<init>", null);
-    TestSnapshotListener listener = installProbes(CLASS_NAME, abstractMethodProbe);
+    TestSnapshotListener listener = installProbes(abstractMethodProbe);
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
     assertNotNull(testClass);
   }
@@ -1635,7 +1575,7 @@ public class CapturedSnapshotTest {
   public void overloadedMethods() throws Exception {
     final String CLASS_NAME = "com.datadog.debugger.CapturedSnapshot16";
     TestSnapshotListener listener =
-        installProbes(CLASS_NAME, createProbe(PROBE_ID, CLASS_NAME, "overload", null));
+        installProbes(createProbe(PROBE_ID, CLASS_NAME, "overload", null));
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
     int result = Reflect.onClass(testClass).call("main", "").get();
     assertEquals(63, result);
@@ -1649,8 +1589,7 @@ public class CapturedSnapshotTest {
   @Test
   public void noDebugInfoEmptyMethod() throws Exception {
     final String CLASS_NAME = "CapturedSnapshot03";
-    TestSnapshotListener listener =
-        installProbes(CLASS_NAME, createProbe(PROBE_ID, CLASS_NAME, "empty", null));
+    TestSnapshotListener listener = installProbes(createProbe(PROBE_ID, CLASS_NAME, "empty", null));
     Map<String, byte[]> classFileBuffers = compile(CLASS_NAME, SourceCompiler.DebugInfo.NONE, "8");
     Class<?> testClass = loadClass(CLASS_NAME, classFileBuffers);
     int result = Reflect.onClass(testClass).call("main", "2").get();
@@ -1703,7 +1642,7 @@ public class CapturedSnapshotTest {
   public void objectDynamicType() throws IOException, URISyntaxException {
     final String CLASS_NAME = "com.datadog.debugger.CapturedSnapshot17";
     TestSnapshotListener listener =
-        installProbes(CLASS_NAME, createProbe(PROBE_ID, CLASS_NAME, "processWithArg", null));
+        installProbes(createProbe(PROBE_ID, CLASS_NAME, "processWithArg", null));
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
     int result = Reflect.onClass(testClass).call("main", "2").get();
     assertEquals(50, result);
@@ -1719,7 +1658,7 @@ public class CapturedSnapshotTest {
   public void exceptionAsLocalVariable() throws IOException, URISyntaxException {
     final String CLASS_NAME = "com.datadog.debugger.CapturedSnapshot18";
     TestSnapshotListener listener =
-        installProbes(CLASS_NAME, createProbe(PROBE_ID, CLASS_NAME, null, null, "14"));
+        installProbes(createProbe(PROBE_ID, CLASS_NAME, null, null, "14"));
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
     int result = Reflect.onClass(testClass).call("main", "2").get();
     assertEquals(42, result);
@@ -1742,7 +1681,7 @@ public class CapturedSnapshotTest {
                 new ProbeCondition(DSL.when(DSL.eq(DSL.ref("arg"), DSL.value("1"))), "arg == '1'"))
             .evaluateAt(MethodLocation.ENTRY)
             .build();
-    TestSnapshotListener listener = installProbes(CLASS_NAME, logProbes);
+    TestSnapshotListener listener = installProbes(logProbes);
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
     int result = Reflect.onClass(testClass).call("main", "1").get();
     assertEquals(3, result);
@@ -1759,7 +1698,7 @@ public class CapturedSnapshotTest {
                     DSL.when(DSL.eq(DSL.ref("@return"), DSL.value(3))), "@return == 3"))
             .evaluateAt(MethodLocation.EXIT)
             .build();
-    TestSnapshotListener listener = installProbes(CLASS_NAME, logProbes);
+    TestSnapshotListener listener = installProbes(logProbes);
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
     int result = Reflect.onClass(testClass).call("main", "1").get();
     assertEquals(3, result);
@@ -1776,7 +1715,7 @@ public class CapturedSnapshotTest {
                     DSL.when(DSL.eq(DSL.ref("@return"), DSL.value(0))), "@return == 0"))
             .evaluateAt(MethodLocation.EXIT)
             .build();
-    TestSnapshotListener listener = installProbes(CLASS_NAME, logProbes);
+    TestSnapshotListener listener = installProbes(logProbes);
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
     int result = Reflect.onClass(testClass).call("main", "1").get();
     assertEquals(3, result);
@@ -1797,7 +1736,7 @@ public class CapturedSnapshotTest {
                 new ProbeCondition(DSL.when(DSL.ge(DSL.ref("after"), DSL.value(0))), "after >= 0"))
             .evaluateAt(MethodLocation.EXIT)
             .build();
-    TestSnapshotListener listener = installProbes(CLASS_NAME, probe);
+    TestSnapshotListener listener = installProbes(probe);
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
     try {
       Reflect.onClass(testClass).call("main", "triggerUncaughtException").get();
@@ -1825,7 +1764,7 @@ public class CapturedSnapshotTest {
   public void uncaughtExceptionCaptureLocalVars() throws IOException, URISyntaxException {
     final String CLASS_NAME = "com.datadog.debugger.CapturedSnapshot31";
     LogProbe probe = createProbeAtExit(PROBE_ID, CLASS_NAME, "uncaughtException", null);
-    TestSnapshotListener listener = installProbes(CLASS_NAME, probe);
+    TestSnapshotListener listener = installProbes(probe);
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
     try {
       Reflect.onClass(testClass).call("main", "uncaughtException").get();
@@ -1849,7 +1788,7 @@ public class CapturedSnapshotTest {
   public void methodProbeLocalVarsLocalScopes() throws IOException, URISyntaxException {
     final String CLASS_NAME = "com.datadog.debugger.CapturedSnapshot31";
     LogProbe probe = createProbeAtExit(PROBE_ID, CLASS_NAME, "localScopes", "(String)");
-    TestSnapshotListener listener = installProbes(CLASS_NAME, probe);
+    TestSnapshotListener listener = installProbes(probe);
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
     int result = Reflect.onClass(testClass).call("main", "localScopes").get();
     assertEquals(42, result);
@@ -1865,7 +1804,7 @@ public class CapturedSnapshotTest {
   public void methodProbeLocalVarsDeepScopes() throws IOException, URISyntaxException {
     final String CLASS_NAME = "com.datadog.debugger.CapturedSnapshot31";
     LogProbe probe = createProbeAtExit(PROBE_ID, CLASS_NAME, "deepScopes", "(String)");
-    TestSnapshotListener listener = installProbes(CLASS_NAME, probe);
+    TestSnapshotListener listener = installProbes(probe);
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
     int result = Reflect.onClass(testClass).call("main", "deepScopes").get();
     assertEquals(4, result);
@@ -1897,7 +1836,7 @@ public class CapturedSnapshotTest {
   public void methodProbeExceptionLocalVars() throws IOException, URISyntaxException {
     final String CLASS_NAME = "com.datadog.debugger.CapturedSnapshot31";
     LogProbe probe = createProbeAtExit(PROBE_ID, CLASS_NAME, "caughtException", "(String)");
-    TestSnapshotListener listener = installProbes(CLASS_NAME, probe);
+    TestSnapshotListener listener = installProbes(probe);
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
     int result = Reflect.onClass(testClass).call("main", "illegalState").get();
     assertEquals(0, result);
@@ -1931,7 +1870,7 @@ public class CapturedSnapshotTest {
   public void overlappingLocalVarSlot() throws IOException, URISyntaxException {
     final String CLASS_NAME = "com.datadog.debugger.CapturedSnapshot31";
     LogProbe probe = createProbeAtExit(PROBE_ID, CLASS_NAME, "overlappingSlots", "(String)");
-    TestSnapshotListener listener = installProbes(CLASS_NAME, probe);
+    TestSnapshotListener listener = installProbes(probe);
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
     int result = Reflect.onClass(testClass).call("main", "overlappingSlots").get();
     assertEquals(5, result);
@@ -1949,7 +1888,7 @@ public class CapturedSnapshotTest {
     final String CLASS_NAME = "com.datadog.debugger.CapturedSnapshot31";
     LogProbe probe =
         createProbeAtExit(PROBE_ID, CLASS_NAME, "duplicateLocalDifferentScope", "(String)");
-    TestSnapshotListener listener = installProbes(CLASS_NAME, probe);
+    TestSnapshotListener listener = installProbes(probe);
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
     int result = Reflect.onClass(testClass).call("main", "duplicateLocalDifferentScope").get();
     assertEquals(28, result);
@@ -1963,7 +1902,7 @@ public class CapturedSnapshotTest {
     final String CLASS_NAME = "com.datadog.debugger.CapturedSnapshot23";
     final String ENUM_CLASS = CLASS_NAME + "$MyEnum";
     TestSnapshotListener listener =
-        installProbes(ENUM_CLASS, createProbe(PROBE_ID, ENUM_CLASS, "<init>", null));
+        installProbes(createProbe(PROBE_ID, ENUM_CLASS, "<init>", null));
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
     int result = Reflect.onClass(testClass).call("main", "").get();
     assertEquals(2, result);
@@ -1980,7 +1919,7 @@ public class CapturedSnapshotTest {
   public void enumValues() throws IOException, URISyntaxException {
     final String CLASS_NAME = "com.datadog.debugger.CapturedSnapshot23";
     TestSnapshotListener listener =
-        installProbes(CLASS_NAME, createProbe(PROBE_ID, CLASS_NAME, "convert", null));
+        installProbes(createProbe(PROBE_ID, CLASS_NAME, "convert", null));
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
     int result = Reflect.onClass(testClass).call("main", "2").get();
     assertEquals(2, result);
@@ -2008,7 +1947,7 @@ public class CapturedSnapshotTest {
                                 DSL.value("com.datadog.debugger.CapturedSnapshot23$MyEnum.TWO")))),
                     "@return == 'TWO' && @return == 'MyEnum.TWO' && @return == 'com.datadog.debugger.CapturedSnapshot23$MyEnum.TWO'"))
             .build();
-    TestSnapshotListener listener = installProbes(CLASS_NAME, probe);
+    TestSnapshotListener listener = installProbes(probe);
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
     int result = Reflect.onClass(testClass).call("main", "2").get();
     assertEquals(2, result);
@@ -2023,8 +1962,7 @@ public class CapturedSnapshotTest {
   public void recursiveCapture() throws IOException, URISyntaxException {
     final String CLASS_NAME = "com.datadog.debugger.CapturedSnapshot24";
     final String INNER_CLASS = CLASS_NAME + "$Holder";
-    TestSnapshotListener listener =
-        installProbes(INNER_CLASS, createProbe(PROBE_ID, INNER_CLASS, "size", null));
+    TestSnapshotListener listener = installProbes(createProbe(PROBE_ID, INNER_CLASS, "size", null));
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
     int result = Reflect.onClass(testClass).call("main", "").get();
     assertEquals(1, result);
@@ -2034,8 +1972,7 @@ public class CapturedSnapshotTest {
   public void recursiveCaptureException() throws IOException, URISyntaxException {
     final String CLASS_NAME = "com.datadog.debugger.CapturedSnapshot24";
     final String INNER_CLASS = CLASS_NAME + "$HolderWithException";
-    TestSnapshotListener listener =
-        installProbes(INNER_CLASS, createProbe(PROBE_ID, INNER_CLASS, "size", null));
+    TestSnapshotListener listener = installProbes(createProbe(PROBE_ID, INNER_CLASS, "size", null));
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
     try {
       Reflect.onClass(testClass).call("main", "exception").get();
@@ -2070,7 +2007,7 @@ public class CapturedSnapshotTest {
                 new ProbeCondition(
                     DSL.when(DSL.ge(DSL.len(DSL.ref("holder")), DSL.value(0))), "len(holder) >= 0"))
             .build();
-    TestSnapshotListener listener = installProbes(CLASS_NAME, logProbe);
+    TestSnapshotListener listener = installProbes(logProbe);
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
     int result = Reflect.onClass(testClass).call("main", "").get();
     Assertions.assertEquals(1, result);
@@ -2105,7 +2042,7 @@ public class CapturedSnapshotTest {
             .template(LOG_TEMPLATE, parseTemplate(LOG_TEMPLATE))
             .evaluateAt(MethodLocation.EXIT)
             .build();
-    TestSnapshotListener listener = installProbes(CLASS_NAME, probe1, probe2);
+    TestSnapshotListener listener = installProbes(probe1, probe2);
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
     int result = Reflect.onClass(testClass).call("main", "1").get();
     assertEquals(3, result);
@@ -2126,7 +2063,7 @@ public class CapturedSnapshotTest {
             .captureSnapshot(true)
             .evaluateAt(MethodLocation.EXIT)
             .build();
-    TestSnapshotListener listener = installProbes(CLASS_NAME, probe1);
+    TestSnapshotListener listener = installProbes(probe1);
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
     int result = Reflect.onClass(testClass).call("main", "secret123").get();
     Assertions.assertEquals(42, result);
@@ -2186,7 +2123,7 @@ public class CapturedSnapshotTest {
             .captureSnapshot(true)
             .evaluateAt(MethodLocation.EXIT)
             .build();
-    TestSnapshotListener listener = installProbes(CLASS_NAME, probe1, probe2, probe3);
+    TestSnapshotListener listener = installProbes(probe1, probe2, probe3);
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
     int result = Reflect.onClass(testClass).call("main", "secret123").get();
     Assertions.assertEquals(42, result);
@@ -2216,7 +2153,7 @@ public class CapturedSnapshotTest {
             .captureSnapshot(true)
             .evaluateAt(MethodLocation.EXIT)
             .build();
-    TestSnapshotListener listener = installProbes(CLASS_NAME, probe1);
+    TestSnapshotListener listener = installProbes(probe1);
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
     int result = Reflect.onClass(testClass).call("main", "secret123").get();
     Assertions.assertEquals(42, result);
@@ -2241,7 +2178,7 @@ public class CapturedSnapshotTest {
             .captureSnapshot(true)
             .evaluateAt(MethodLocation.EXIT)
             .build();
-    TestSnapshotListener listener = installProbes(CLASS_NAME, probe1);
+    TestSnapshotListener listener = installProbes(probe1);
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
     int result = Reflect.onClass(testClass).call("main", "secret123").get();
     Assertions.assertEquals(42, result);
@@ -2304,7 +2241,7 @@ public class CapturedSnapshotTest {
             .captureSnapshot(true)
             .evaluateAt(MethodLocation.EXIT)
             .build();
-    TestSnapshotListener listener = installProbes(CLASS_NAME, probe1, probe2, probe3);
+    TestSnapshotListener listener = installProbes(probe1, probe2, probe3);
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
     int result = Reflect.onClass(testClass).call("main", "secret123").get();
     Assertions.assertEquals(42, result);
@@ -2354,10 +2291,6 @@ public class CapturedSnapshotTest {
     doSamplingTest(this::lineProbeCondition, 1, 1);
   }
 
-  interface TestMethod {
-    void run() throws IOException, URISyntaxException;
-  }
-
   private void doSamplingTest(TestMethod testRun, int expectedGlobalCount, int expectedProbeCount)
       throws IOException, URISyntaxException {
     MockSampler probeSampler = new MockSampler();
@@ -2379,7 +2312,7 @@ public class CapturedSnapshotTest {
     final String CLASS_NAME = "com.datadog.debugger.CapturedSnapshot29";
     final String RECORD_NAME = "com.datadog.debugger.MyRecord1";
     LogProbe probe1 = createProbeAtExit(PROBE_ID, RECORD_NAME, "age", null);
-    TestSnapshotListener listener = installProbes(RECORD_NAME, probe1);
+    TestSnapshotListener listener = installProbes(probe1);
     Class<?> testClass = compileAndLoadClass(CLASS_NAME, "17");
     int result = Reflect.onClass(testClass).call("main", "1").get();
     assertEquals(42, result);
@@ -2399,7 +2332,7 @@ public class CapturedSnapshotTest {
     final String CLASS_NAME = "com.datadog.debugger.CapturedSnapshot29";
     final String RECORD_NAME = "com.datadog.debugger.MyRecord2";
     LogProbe probe1 = createProbe(PROBE_ID, RECORD_NAME, null, null, "29");
-    TestSnapshotListener listener = installProbes(RECORD_NAME, probe1);
+    TestSnapshotListener listener = installProbes(probe1);
     Class<?> testClass = compileAndLoadClass(CLASS_NAME, "17");
     int result = Reflect.onClass(testClass).call("main", "1").get();
     assertEquals(42, result);
@@ -2451,7 +2384,7 @@ public class CapturedSnapshotTest {
     TestTraceInterceptor traceInterceptor = new TestTraceInterceptor();
     tracer.addTraceInterceptor(traceInterceptor);
     try {
-      TestSnapshotListener snapshotListener = installProbes(CLASS_NAME, configuration);
+      TestSnapshotListener snapshotListener = installProbes(configuration);
       DebuggerContext.initTracer(new DebuggerTracer(mock(ProbeStatusSink.class)));
       MetricProbesInstrumentationTest.MetricForwarderListener metricListener =
           new MetricProbesInstrumentationTest.MetricForwarderListener();
@@ -2491,7 +2424,7 @@ public class CapturedSnapshotTest {
                 "dd_watches_dsl:{typed.fld.fld.msg},{nullTyped.fld}",
                 "dd_watches_json:[{\"dsl\":\"typed.fld.fld.msg_json\",\"json\":{\"getmember\":[{\"getmember\":[{\"getmember\":[{\"ref\":\"typed\"},\"fld\"]},\"fld\"]},\"msg\"]}}]")
             .build();
-    TestSnapshotListener listener = installProbes(CLASS_NAME, probe);
+    TestSnapshotListener listener = installProbes(probe);
     Class<?> testClass = compileAndLoadClass(CLASS_NAME);
     int result = Reflect.onClass(testClass).call("main", "1").get();
     assertEquals(3, result);
@@ -2555,384 +2488,5 @@ public class CapturedSnapshotTest {
     Snapshot snapshot = listener.snapshots.get(0);
     assertEquals(PROBE_ID.getId(), snapshot.getProbe().getId());
     return snapshot;
-  }
-
-  private TestSnapshotListener installSingleProbe(
-      String typeName, String methodName, String signature, String... lines) {
-    LogProbe logProbes = createProbe(PROBE_ID, typeName, methodName, signature, lines);
-    return installProbes(typeName, logProbes);
-  }
-
-  private TestSnapshotListener installSingleProbeAtExit(
-      String typeName, String methodName, String signature, String... lines) {
-    LogProbe logProbes = createProbeAtExit(PROBE_ID, typeName, methodName, signature, lines);
-    return installProbes(typeName, logProbes);
-  }
-
-  private TestSnapshotListener installProbes(
-      String expectedClassName, Configuration configuration) {
-    Config config = mock(Config.class);
-    when(config.isDebuggerEnabled()).thenReturn(true);
-    when(config.isDebuggerClassFileDumpEnabled()).thenReturn(true);
-    when(config.isDebuggerVerifyByteCode()).thenReturn(false);
-    when(config.getFinalDebuggerSnapshotUrl())
-        .thenReturn("http://localhost:8126/debugger/v1/input");
-    when(config.getFinalDebuggerSymDBUrl()).thenReturn("http://localhost:8126/symdb/v1/input");
-    Collection<LogProbe> logProbes = configuration.getLogProbes();
-    instrumentationListener = new MockInstrumentationListener();
-    probeStatusSink = mock(ProbeStatusSink.class);
-    TestSnapshotListener listener = new TestSnapshotListener(config, probeStatusSink);
-    currentTransformer =
-        new DebuggerTransformer(config, configuration, instrumentationListener, listener);
-    instr.addTransformer(currentTransformer);
-    DebuggerAgentHelper.injectSink(listener);
-    DebuggerContext.initProbeResolver(
-        (encodedId) ->
-            resolver(
-                encodedId,
-                logProbes,
-                configuration.getSpanDecorationProbes(),
-                configuration.getTriggerProbes()));
-    DebuggerContext.initClassFilter(new DenyListHelper(null));
-    DebuggerContext.initValueSerializer(new JsonSnapshotSerializer());
-    if (logProbes != null) {
-      for (LogProbe probe : logProbes) {
-        if (probe.getSampling() != null) {
-          ProbeRateLimiter.setRate(
-              probe.getId(),
-              probe.getSampling().getSnapshotsPerSecond(),
-              probe.isCaptureSnapshot());
-        }
-      }
-    }
-    if (configuration.getSampling() != null) {
-      ProbeRateLimiter.setGlobalSnapshotRate(configuration.getSampling().getSnapshotsPerSecond());
-    }
-    return listener;
-  }
-
-  private ProbeImplementation resolver(
-      String encodedId,
-      Collection<LogProbe> logProbes,
-      Collection<SpanDecorationProbe> spanDecorationProbes,
-      Collection<TriggerProbe> triggerProbes) {
-    for (LogProbe probe : logProbes) {
-      if (probe.getProbeId().getEncodedId().equals(encodedId)) {
-        return probe;
-      }
-    }
-    for (SpanDecorationProbe probe : spanDecorationProbes) {
-      if (probe.getProbeId().getEncodedId().equals(encodedId)) {
-        return probe;
-      }
-    }
-    for (TriggerProbe probe : triggerProbes) {
-      if (probe.getProbeId().getEncodedId().equals(encodedId)) {
-        return probe;
-      }
-    }
-    return null;
-  }
-
-  private TestSnapshotListener installProbes(String expectedClassName, LogProbe... logProbes) {
-    return installProbes(
-        expectedClassName,
-        Configuration.builder()
-            .setService(SERVICE_NAME)
-            .addLogProbes(Arrays.asList(logProbes))
-            .build());
-  }
-
-  private void assertCaptureArgs(
-      CapturedContext context, String name, String typeName, String value) {
-    CapturedContext.CapturedValue capturedValue = context.getArguments().get(name);
-    assertEquals(typeName, capturedValue.getType());
-    assertEquals(value, MoshiSnapshotTestHelper.getValue(capturedValue));
-  }
-
-  private void assertCaptureLocals(
-      CapturedContext context, String name, String typeName, String value) {
-    CapturedContext.CapturedValue localVar = context.getLocals().get(name);
-    assertEquals(typeName, localVar.getType());
-    assertEquals(value, MoshiSnapshotTestHelper.getValue(localVar));
-  }
-
-  private void assertCaptureLocals(
-      CapturedContext context, String name, String typeName, Map<String, String> expectedFields) {
-    CapturedContext.CapturedValue localVar = context.getLocals().get(name);
-    assertEquals(typeName, localVar.getType());
-    Map<String, CapturedContext.CapturedValue> fields = getFields(localVar);
-    for (Map.Entry<String, String> entry : expectedFields.entrySet()) {
-      assertTrue(fields.containsKey(entry.getKey()));
-      CapturedContext.CapturedValue fieldCapturedValue = fields.get(entry.getKey());
-      if (fieldCapturedValue.getNotCapturedReason() != null) {
-        assertEquals(entry.getValue(), String.valueOf(fieldCapturedValue.getNotCapturedReason()));
-      } else {
-        assertEquals(entry.getValue(), String.valueOf(fieldCapturedValue.getValue()));
-      }
-    }
-  }
-
-  private void assertCaptureFields(
-      CapturedContext context, String name, String typeName, String value) {
-    CapturedContext.CapturedValue field = getFields(context.getArguments().get("this")).get(name);
-    assertEquals(typeName, field.getType());
-    assertEquals(value, MoshiSnapshotTestHelper.getValue(field));
-  }
-
-  private void assertCaptureFields(
-      CapturedContext context, String name, String typeName, Collection<?> collection) {
-    CapturedContext.CapturedValue field = getFields(context.getArguments().get("this")).get(name);
-    assertEquals(typeName, field.getType());
-    Iterator<?> iterator = collection.iterator();
-    for (Object obj : getCollection(field)) {
-      if (iterator.hasNext()) {
-        assertEquals(iterator.next(), obj);
-      } else {
-        Assertions.fail("not same number of elements");
-      }
-    }
-  }
-
-  private void assertCaptureFields(
-      CapturedContext context, String name, String typeName, Map<Object, Object> expectedMap) {
-    CapturedContext.CapturedValue field = getFields(context.getArguments().get("this")).get(name);
-    assertEquals(typeName, field.getType());
-    Map<Object, Object> map = getMap(field);
-    assertEquals(expectedMap.size(), map.size());
-    for (Map.Entry<Object, Object> entry : map.entrySet()) {
-      assertTrue(expectedMap.containsKey(entry.getKey()));
-      assertEquals(expectedMap.get(entry.getKey()), entry.getValue());
-    }
-  }
-
-  private void assertCaptureFieldsNotCaptured(
-      CapturedContext context, String name, String expectedReasonRegEx) {
-    CapturedContext.CapturedValue field = getFields(context.getArguments().get("this")).get(name);
-    assertTrue(
-        field.getNotCapturedReason().matches(expectedReasonRegEx), field.getNotCapturedReason());
-  }
-
-  private void assertCaptureStaticFieldsNotCaptured(
-      CapturedContext context, String name, String expectedReasonRegEx) {
-    CapturedContext.CapturedValue field = context.getStaticFields().get(name);
-    assertTrue(
-        field.getNotCapturedReason().matches(expectedReasonRegEx), field.getNotCapturedReason());
-  }
-
-  private void assertCaptureFieldCount(CapturedContext context, int expectedFieldCount) {
-    assertEquals(expectedFieldCount, getFields(context.getArguments().get("this")).size());
-  }
-
-  private void assertCaptureReturnValue(CapturedContext context, String typeName, String value) {
-    CapturedContext.CapturedValue returnValue = context.getLocals().get("@return");
-    assertEquals(typeName, returnValue.getType());
-    assertEquals(value, MoshiSnapshotTestHelper.getValue(returnValue));
-  }
-
-  private void assertCaptureReturnValue(
-      CapturedContext context, String typeName, Map<String, String> expectedFields) {
-    CapturedContext.CapturedValue returnValue = context.getLocals().get("@return");
-    assertEquals(typeName, returnValue.getType());
-    Map<String, CapturedContext.CapturedValue> fields = getFields(returnValue);
-    for (Map.Entry<String, String> entry : expectedFields.entrySet()) {
-      assertTrue(fields.containsKey(entry.getKey()));
-      CapturedContext.CapturedValue fieldCapturedValue = fields.get(entry.getKey());
-      if (fieldCapturedValue.getNotCapturedReason() != null) {
-        assertEquals(entry.getValue(), String.valueOf(fieldCapturedValue.getNotCapturedReason()));
-      } else {
-        assertEquals(entry.getValue(), String.valueOf(fieldCapturedValue.getValue()));
-      }
-    }
-  }
-
-  private void assertCaptureWatches(
-      CapturedContext context, String name, String typeName, String value) {
-    CapturedContext.CapturedValue watch = context.getWatches().get(name);
-    assertEquals(typeName, watch.getType());
-    assertEquals(value, MoshiSnapshotTestHelper.getValue(watch));
-  }
-
-  private void assertCaptureThrowable(
-      CapturedContext context, String typeName, String message, String methodName, int lineNumber) {
-    CapturedContext.CapturedThrowable throwable = context.getCapturedThrowable();
-    assertCaptureThrowable(throwable, typeName, message, methodName, lineNumber);
-  }
-
-  private void assertCaptureThrowable(
-      CapturedContext.CapturedThrowable throwable,
-      String typeName,
-      String message,
-      String methodName,
-      int lineNumber) {
-    assertNotNull(throwable);
-    assertEquals(typeName, throwable.getType());
-    assertEquals(message, throwable.getMessage());
-    assertNotNull(throwable.getStacktrace());
-    Assertions.assertFalse(throwable.getStacktrace().isEmpty());
-    assertEquals(methodName, throwable.getStacktrace().get(0).getFunction());
-    assertEquals(lineNumber, throwable.getStacktrace().get(0).getLineNumber());
-  }
-
-  public static Map<String, CapturedContext.CapturedValue> getFields(
-      CapturedContext.CapturedValue capturedValue) {
-    try {
-      CapturedContext.CapturedValue valued = VALUE_ADAPTER.fromJson(capturedValue.getStrValue());
-      Map<String, CapturedContext.CapturedValue> results = new HashMap<>();
-      if (valued.getNotCapturedReason() != null) {
-        results.put(
-            "@" + NOT_CAPTURED_REASON,
-            CapturedContext.CapturedValue.notCapturedReason(
-                null, null, valued.getNotCapturedReason()));
-      }
-      if (valued.getValue() == null) {
-        return results;
-      }
-      results.putAll((Map<String, CapturedContext.CapturedValue>) valued.getValue());
-      return results;
-    } catch (IOException e) {
-      e.printStackTrace();
-      return Collections.emptyMap();
-    }
-  }
-
-  private static Collection<?> getCollection(CapturedContext.CapturedValue capturedValue) {
-    try {
-      if (capturedValue.getStrValue() == null) {
-        return (Collection<?>) capturedValue.getValue();
-      }
-      Map<String, Object> capturedValueMap = GENERIC_ADAPTER.fromJson(capturedValue.getStrValue());
-      List<Object> elements = (List<Object>) capturedValueMap.get("elements");
-      if (elements == null) {
-        Assertions.fail("not a collection");
-      }
-      List<Object> result = new ArrayList<>();
-      for (Object obj : elements) {
-        Map<String, Object> element = (Map<String, Object>) obj;
-        String type = (String) element.get("type");
-        if (type == null) {
-          Assertions.fail("no type for element");
-        }
-        if (SerializerWithLimits.isPrimitive(type)) {
-          result.add(element.get("value"));
-        } else {
-          Assertions.fail("not implemented");
-        }
-      }
-      return result;
-    } catch (IOException e) {
-      e.printStackTrace();
-      return null;
-    }
-  }
-
-  private Map<Object, Object> getMap(CapturedContext.CapturedValue capturedValue) {
-    try {
-      if (capturedValue.getStrValue() == null) {
-        return (Map<Object, Object>) capturedValue.getValue();
-      }
-      Map<String, Object> capturedValueMap = GENERIC_ADAPTER.fromJson(capturedValue.getStrValue());
-      List<Object> entries = (List<Object>) capturedValueMap.get("entries");
-      if (entries == null) {
-        Assertions.fail("not a map");
-      }
-      Map<Object, Object> result = new HashMap<>();
-      for (Object obj : entries) {
-        List<Object> entry = (List<Object>) obj;
-        Map<String, Object> key = (Map<String, Object>) entry.get(0);
-        Map<String, Object> value = (Map<String, Object>) entry.get(1);
-        result.put(key.get("value"), value.get("value"));
-      }
-      return result;
-    } catch (IOException e) {
-      e.printStackTrace();
-      return null;
-    }
-  }
-
-  private static LogProbe createProbe(
-      ProbeId id, String typeName, String methodName, String signature, String... lines) {
-    return createProbeBuilder(id, typeName, methodName, signature, lines).build();
-  }
-
-  private static LogProbe createProbeAtExit(
-      ProbeId id, String typeName, String methodName, String signature, String... lines) {
-    return createProbeBuilder(id, typeName, methodName, signature, lines)
-        .evaluateAt(MethodLocation.EXIT)
-        .build();
-  }
-
-  private static LogProbe.Builder createProbeBuilder(
-      ProbeId id, String typeName, String methodName, String signature, String... lines) {
-    return LogProbe.builder()
-        .language(LANGUAGE)
-        .probeId(id)
-        .captureSnapshot(true)
-        .where(typeName, methodName, signature, lines)
-        // Increase sampling limit to avoid being sampled during tests
-        .sampling(new LogProbe.Sampling(100));
-  }
-
-  private static LogProbe createSourceFileProbe(ProbeId id, String sourceFile, int line) {
-    return new LogProbe.Builder()
-        .language(LANGUAGE)
-        .probeId(id)
-        .captureSnapshot(true)
-        .where(null, null, null, line, sourceFile)
-        .evaluateAt(MethodLocation.EXIT)
-        .build();
-  }
-
-  static class MockInstrumentationListener implements DebuggerTransformer.InstrumentationListener {
-    final Map<String, InstrumentationResult> results = new HashMap<>();
-
-    @Override
-    public void instrumentationResult(ProbeDefinition definition, InstrumentationResult result) {
-      results.put(definition.getId(), result);
-    }
-  }
-
-  static class KotlinHelper {
-    public static Class<?> compileAndLoad(
-        String className, String sourceFileName, List<File> outputFilesToDelete) {
-      K2JVMCompiler compiler = new K2JVMCompiler();
-      K2JVMCompilerArguments args = compiler.createArguments();
-      args.setFreeArgs(Collections.singletonList(sourceFileName));
-      String compilerOutputDir = "/tmp/" + CapturedSnapshotTest.class.getSimpleName() + "-kotlin";
-      args.setDestination(compilerOutputDir);
-      args.setClasspath(System.getProperty("java.class.path"));
-      ExitCode exitCode =
-          compiler.execImpl(
-              new PrintingMessageCollector(System.out, MessageRenderer.WITHOUT_PATHS, true),
-              Services.EMPTY,
-              args);
-      if (exitCode.getCode() != 0) {
-        throw new RuntimeException("Kotlin compilation failed");
-      }
-      File compileOutputDirFile = new File(compilerOutputDir);
-      try {
-        URLClassLoader urlClassLoader =
-            new URLClassLoader(new URL[] {compileOutputDirFile.toURI().toURL()});
-        return urlClassLoader.loadClass(className);
-      } catch (Exception ex) {
-        throw new RuntimeException(ex);
-      } finally {
-        registerFilesToDeleteDir(compileOutputDirFile, outputFilesToDelete);
-      }
-    }
-
-    public static void registerFilesToDeleteDir(File dir, List<File> outputFilesToDelete) {
-      if (!dir.exists()) {
-        return;
-      }
-      try {
-        Files.walk(dir.toPath())
-            .sorted(Comparator.reverseOrder())
-            .map(Path::toFile)
-            .forEach(outputFilesToDelete::add);
-      } catch (IOException ex) {
-        ex.printStackTrace();
-      }
-    }
   }
 }
