+++ dd-java-agent/agent-debugger/src/test/java/com/datadog/debugger/util/TestHelper.java
@@ -2,7 +2,6 @@ package com.datadog.debugger.util;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-import datadog.trace.api.Config;
 import java.lang.reflect.Field;
 import java.time.Duration;
 import java.util.Map;
@@ -10,11 +9,11 @@ import java.util.function.BooleanSupplier;
 
 public class TestHelper {
 
-  public static void setFieldInConfig(Config config, String fieldName, Object value) {
+  public static void setFieldInConfig(Object target, String fieldName, Object value) {
     try {
-      Field field = config.getClass().getDeclaredField(fieldName);
+      Field field = target.getClass().getDeclaredField(fieldName);
       field.setAccessible(true);
-      field.set(config, value);
+      field.set(target, value);
     } catch (Throwable e) {
       e.printStackTrace();
     }
